app/requirements.txt
fastapi>=0.79.0,<0.80.0
pydantic>=1.9.0,<2.0.0
uvicorn>=0.18.0,<0.19.0
jinja2>=3.1.0,<3.2.0
markdown2>=2.4.0,<2.5.0
requests>=2.28.0,<2.29.0